integration/pyproject.toml
[project]
name = "integration"
version = "0.1.0"
description = "Ares integration adapter — bridges muna-agentsdk contracts to Hermes pentest agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "muna-agentsdk",
    "muna-vaultsdk",
    "nats-py>=2.14.0",
    "uvicorn>=0.46.0",
]

[tool.uv.sources]
muna-agentsdk = { path = "../../muna-agentsdk", editable = true }
muna-vaultsdk = { path = "../../muna-vaultsdk", editable = true }

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["integration: requires NATS and Hermes"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
